+++ TencentCloud/Ess/V20201111/Models/CreateSchemeUrlResponse.cs
@@ -30,6 +30,9 @@ namespace TencentCloud.Ess.V20201111.Models
         /// <ul><li>如果EndPoint是**APP**，得到的链接类似于`pages/guide?from=default&where=mini&id=yDwJSUUirqauh***7jNSxwdirTSGuH&to=CONTRACT_DETAIL&name=&phone=&shortKey=yDw***k1xFc5`, 用法可以参加接口描述中的"跳转到小程序的实现"</li>
         /// <li>如果EndPoint是**HTTP**，得到的链接类似于 `https://res.ess.tencent.cn/cdn/h5-activity/jump-mp.html?where=mini&from=SFY&id=yDwfEUUw**4rV6Avz&to=MVP_CONTRACT_COVER&name=%E9%83%**5%86%9B`，点击后会跳转到腾讯电子签小程序进行签署</li>
         /// <li>如果EndPoint是**HTTP_SHORT_URL**，得到的链接类似于 `https://essurl.cn/2n**42Nd`，点击后会跳转到腾讯电子签小程序进行签署</li></ul>
+        /// 
+        /// 
+        /// 注： <font color="red">生成的链路后面不能再增加参数</font>
         /// </summary>
         [JsonProperty("SchemeUrl")]
         public string SchemeUrl{ get; set; }