+++ TencentCloud/Ess/V20201111/Models/FlowApproverUrlInfo.cs
@@ -25,11 +25,11 @@ namespace TencentCloud.Ess.V20201111.Models
     {
         
         /// <summary>
-        /// 签署短链接。</br>
+        /// 签署短链接。
         /// 注意:
-        /// - 该链接有效期为**30分钟**，同时需要注意保密，不要外泄给无关用户。
-        /// - 该链接不支持小程序嵌入，仅支持**移动端浏览器**打开。
-        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// 1. 该链接有效期为<b>30分钟</b>，同时需要注意保密，不要外泄给无关用户。
+        /// 2. 该链接不支持小程序嵌入，仅支持<b>移动端浏览器</b>打开。
+        /// 3. <font color="red">生成的链路后面不能再增加参数</font>（会出现覆盖链接中已有参数导致错误）
         /// </summary>
         [JsonProperty("SignUrl")]
         public string SignUrl{ get; set; }
@@ -37,31 +37,28 @@ namespace TencentCloud.Ess.V20201111.Models
         /// <summary>
         /// 签署人类型。
         /// - **1**: 个人
-        /// 注意：此字段可能返回 null，表示取不到有效值。
         /// </summary>
         [JsonProperty("ApproverType")]
         public long? ApproverType{ get; set; }
 
         /// <summary>
         /// 签署人姓名。
-        /// 注意：此字段可能返回 null，表示取不到有效值。
         /// </summary>
         [JsonProperty("ApproverName")]
         public string ApproverName{ get; set; }
 
         /// <summary>
         /// 签署人手机号。
-        /// 注意：此字段可能返回 null，表示取不到有效值。
         /// </summary>
         [JsonProperty("ApproverMobile")]
         public string ApproverMobile{ get; set; }
 
         /// <summary>
-        /// 签署长链接。</br>
+        /// 签署长链接。
         /// 注意:
-        /// - 该链接有效期为**30分钟**，同时需要注意保密，不要外泄给无关用户。
-        /// - 该链接不支持小程序嵌入，仅支持**移动端浏览器**打开。
-        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// 1. 该链接有效期为**30分钟**，同时需要注意保密，不要外泄给无关用户。
+        /// 2. 该链接不支持小程序嵌入，仅支持**移动端浏览器**打开。
+        /// 3. <font color="red">生成的链路后面不能再增加参数</font>（会出现覆盖链接中已有参数导致错误）
         /// </summary>
         [JsonProperty("LongUrl")]
         public string LongUrl{ get; set; }