+++ TencentCloud/Essbasic/V20210526/Models/SignUrlInfo.cs
@@ -26,6 +26,8 @@ namespace TencentCloud.Essbasic.V20210526.Models
         
         /// <summary>
         /// 签署链接，过期时间为90天
+        /// 
+        /// 注：<font color="red">生成的链路后面不能再增加参数</font>（会出现覆盖链接中已有参数导致错误）
         /// 注意：此字段可能返回 null，表示取不到有效值。
         /// </summary>
         [JsonProperty("SignUrl")]