+++ TencentCloud/Omics/V20221128/Models/RetryRunsResponse.cs
@@ -25,6 +25,13 @@ namespace TencentCloud.Omics.V20221128.Models
     {
         
         /// <summary>
+        /// 新的任务批次ID。
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("RunGroupId")]
+        public string RunGroupId{ get; set; }
+
+        /// <summary>
         /// 唯一请求 ID，每次请求都会返回。定位问题时需要提供该次请求的 RequestId。
         /// </summary>
         [JsonProperty("RequestId")]
@@ -36,6 +43,7 @@ namespace TencentCloud.Omics.V20221128.Models
         /// </summary>
         public override void ToMap(Dictionary<string, string> map, string prefix)
         {
+            this.SetParamSimple(map, prefix + "RunGroupId", this.RunGroupId);
             this.SetParamSimple(map, prefix + "RequestId", this.RequestId);
         }
     }