+++ TencentCloud/Ssm/V20190923/Models/CreateSecretRequest.cs
@@ -55,13 +55,13 @@ namespace TencentCloud.Ssm.V20190923.Models
         public ulong? SecretType{ get; set; }
 
         /// <summary>
-        /// 二进制凭据信息base64编码后的明文。SecretBinary 和 SecretString 必须且只能设置一个，最大支持4096字节。
+        /// 二进制凭据信息base64编码后的明文。SecretBinary 和 SecretString 必须且只能设置一个，最大支持32KB字节。
         /// </summary>
         [JsonProperty("SecretBinary")]
         public string SecretBinary{ get; set; }
 
         /// <summary>
-        /// 文本类型凭据信息明文（不需要进行base64编码）。SecretBinary 和 SecretString 必须且只能设置一个，，最大支持4096字节。
+        /// 文本类型凭据信息明文（不需要进行base64编码）。SecretBinary 和 SecretString 必须且只能设置一个，最大支持32KB字节。
         /// </summary>
         [JsonProperty("SecretString")]
         public string SecretString{ get; set; }