+++ TencentCloud/Wedata/V20210820/Models/ColumnBasicInfo.cs
@@ -1,0 +1,188 @@
+/*
+ * Copyright (c) 2018 THL A29 Limited, a Tencent company. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+namespace TencentCloud.Wedata.V20210820.Models
+{
+    using Newtonsoft.Json;
+    using System.Collections.Generic;
+    using TencentCloud.Common;
+
+    public class ColumnBasicInfo : AbstractModel
+    {
+        
+        /// <summary>
+        /// 表的全局唯一ID
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("TableId")]
+        public string TableId{ get; set; }
+
+        /// <summary>
+        /// 数据源全局唯一ID
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("DatasourceId")]
+        public string DatasourceId{ get; set; }
+
+        /// <summary>
+        /// 数据源名
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("DatasourceName")]
+        public string DatasourceName{ get; set; }
+
+        /// <summary>
+        /// 数据库ID
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("DatabaseId")]
+        public string DatabaseId{ get; set; }
+
+        /// <summary>
+        /// 数据库名
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("DatabaseName")]
+        public string DatabaseName{ get; set; }
+
+        /// <summary>
+        /// 表名称
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("TableName")]
+        public string TableName{ get; set; }
+
+        /// <summary>
+        /// 字段名称
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("ColumnName")]
+        public string ColumnName{ get; set; }
+
+        /// <summary>
+        /// 数据类型,string/int等
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("DataType")]
+        public string DataType{ get; set; }
+
+        /// <summary>
+        /// 字段类型, varchar(32)/int(10)等
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("ColumnType")]
+        public string ColumnType{ get; set; }
+
+        /// <summary>
+        /// 字段默认值
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("ColumnDefault")]
+        public string ColumnDefault{ get; set; }
+
+        /// <summary>
+        /// 索引类型, PRI/MUL/PARTITION等,普通字段该值为空串
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("ColumnKey")]
+        public string ColumnKey{ get; set; }
+
+        /// <summary>
+        /// 字段顺序标识
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("ColumnPosition")]
+        public long? ColumnPosition{ get; set; }
+
+        /// <summary>
+        /// 字段注释
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("ColumnComment")]
+        public string ColumnComment{ get; set; }
+
+        /// <summary>
+        /// 数据类型
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("StoreType")]
+        public string StoreType{ get; set; }
+
+        /// <summary>
+        /// 所属项目ID
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("ProjectId")]
+        public string ProjectId{ get; set; }
+
+        /// <summary>
+        /// 所属项目英文名
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("ProjectName")]
+        public string ProjectName{ get; set; }
+
+        /// <summary>
+        /// 所属项目中文名
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("ProjectDisplayName")]
+        public string ProjectDisplayName{ get; set; }
+
+        /// <summary>
+        /// 创建时间
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("CreateTime")]
+        public string CreateTime{ get; set; }
+
+        /// <summary>
+        /// 更新时间
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("UpdateTime")]
+        public string UpdateTime{ get; set; }
+
+
+        /// <summary>
+        /// For internal usage only. DO NOT USE IT.
+        /// </summary>
+        public override void ToMap(Dictionary<string, string> map, string prefix)
+        {
+            this.SetParamSimple(map, prefix + "TableId", this.TableId);
+            this.SetParamSimple(map, prefix + "DatasourceId", this.DatasourceId);
+            this.SetParamSimple(map, prefix + "DatasourceName", this.DatasourceName);
+            this.SetParamSimple(map, prefix + "DatabaseId", this.DatabaseId);
+            this.SetParamSimple(map, prefix + "DatabaseName", this.DatabaseName);
+            this.SetParamSimple(map, prefix + "TableName", this.TableName);
+            this.SetParamSimple(map, prefix + "ColumnName", this.ColumnName);
+            this.SetParamSimple(map, prefix + "DataType", this.DataType);
+            this.SetParamSimple(map, prefix + "ColumnType", this.ColumnType);
+            this.SetParamSimple(map, prefix + "ColumnDefault", this.ColumnDefault);
+            this.SetParamSimple(map, prefix + "ColumnKey", this.ColumnKey);
+            this.SetParamSimple(map, prefix + "ColumnPosition", this.ColumnPosition);
+            this.SetParamSimple(map, prefix + "ColumnComment", this.ColumnComment);
+            this.SetParamSimple(map, prefix + "StoreType", this.StoreType);
+            this.SetParamSimple(map, prefix + "ProjectId", this.ProjectId);
+            this.SetParamSimple(map, prefix + "ProjectName", this.ProjectName);
+            this.SetParamSimple(map, prefix + "ProjectDisplayName", this.ProjectDisplayName);
+            this.SetParamSimple(map, prefix + "CreateTime", this.CreateTime);
+            this.SetParamSimple(map, prefix + "UpdateTime", this.UpdateTime);
+        }
+    }
+}
+