+++ TencentCloud/Wedata/V20210820/Models/TableBasicInfo.cs
@@ -1,0 +1,260 @@
+/*
+ * Copyright (c) 2018 THL A29 Limited, a Tencent company. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+namespace TencentCloud.Wedata.V20210820.Models
+{
+    using Newtonsoft.Json;
+    using System.Collections.Generic;
+    using TencentCloud.Common;
+
+    public class TableBasicInfo : AbstractModel
+    {
+        
+        /// <summary>
+        /// 表的全局唯一ID
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("TableId")]
+        public string TableId{ get; set; }
+
+        /// <summary>
+        /// 数据源全局唯一ID
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("DatasourceId")]
+        public string DatasourceId{ get; set; }
+
+        /// <summary>
+        /// 数据源名
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("DatasourceName")]
+        public string DatasourceName{ get; set; }
+
+        /// <summary>
+        /// 数据库ID
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("DatabaseId")]
+        public string DatabaseId{ get; set; }
+
+        /// <summary>
+        /// 数据库名
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("DatabaseName")]
+        public string DatabaseName{ get; set; }
+
+        /// <summary>
+        /// 表名称
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("TableName")]
+        public string TableName{ get; set; }
+
+        /// <summary>
+        /// 引擎/存储类型
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("EngineType")]
+        public string EngineType{ get; set; }
+
+        /// <summary>
+        /// 表类型，视图，外部表等
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("TableType")]
+        public string TableType{ get; set; }
+
+        /// <summary>
+        /// 项目Id
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("ProjectId")]
+        public string ProjectId{ get; set; }
+
+        /// <summary>
+        /// 所属项目英文名
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("ProjectName")]
+        public string ProjectName{ get; set; }
+
+        /// <summary>
+        /// 所属项目英中文名
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("ProjectDisplayName")]
+        public string ProjectDisplayName{ get; set; }
+
+        /// <summary>
+        /// 责任人ID
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("TableOwnerId")]
+        public string TableOwnerId{ get; set; }
+
+        /// <summary>
+        /// 责任人名称
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("TableOwnerName")]
+        public string TableOwnerName{ get; set; }
+
+        /// <summary>
+        /// 存储位置
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("StorageLocation")]
+        public long? StorageLocation{ get; set; }
+
+        /// <summary>
+        /// 表描述
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("Description")]
+        public string Description{ get; set; }
+
+        /// <summary>
+        /// 是否分区表，0-全量表 1-分区表
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("IsPartitionTable")]
+        public long? IsPartitionTable{ get; set; }
+
+        /// <summary>
+        /// 分区字段list
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("PartitionColumns")]
+        public string[] PartitionColumns{ get; set; }
+
+        /// <summary>
+        /// 存储格式
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("StorageFormat")]
+        public string StorageFormat{ get; set; }
+
+        /// <summary>
+        /// 存储量，字节数
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("StorageSize")]
+        public ulong? StorageSize{ get; set; }
+
+        /// <summary>
+        /// 存储量，单位
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("StorageSizeWithUnit")]
+        public string StorageSizeWithUnit{ get; set; }
+
+        /// <summary>
+        /// 累计存储【MB】
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("TotalSizeMb")]
+        public ulong? TotalSizeMb{ get; set; }
+
+        /// <summary>
+        /// 副本数量
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("ReplicaCount")]
+        public long? ReplicaCount{ get; set; }
+
+        /// <summary>
+        /// 文件数
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("FileCount")]
+        public long? FileCount{ get; set; }
+
+        /// <summary>
+        /// 分区总数（包含hive，iceberg）
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("PartitionCount")]
+        public long? PartitionCount{ get; set; }
+
+        /// <summary>
+        /// 分区字段数量（包含hive，iceberg）
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("PartitionFieldCount")]
+        public long? PartitionFieldCount{ get; set; }
+
+        /// <summary>
+        /// 生命周期-分区保留天数【分区保留策略时有效】
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("PartitionExpireDays")]
+        public long? PartitionExpireDays{ get; set; }
+
+        /// <summary>
+        /// 创建时间
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("CreateTime")]
+        public string CreateTime{ get; set; }
+
+        /// <summary>
+        /// 更新时间
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("UpdateTime")]
+        public string UpdateTime{ get; set; }
+
+
+        /// <summary>
+        /// For internal usage only. DO NOT USE IT.
+        /// </summary>
+        public override void ToMap(Dictionary<string, string> map, string prefix)
+        {
+            this.SetParamSimple(map, prefix + "TableId", this.TableId);
+            this.SetParamSimple(map, prefix + "DatasourceId", this.DatasourceId);
+            this.SetParamSimple(map, prefix + "DatasourceName", this.DatasourceName);
+            this.SetParamSimple(map, prefix + "DatabaseId", this.DatabaseId);
+            this.SetParamSimple(map, prefix + "DatabaseName", this.DatabaseName);
+            this.SetParamSimple(map, prefix + "TableName", this.TableName);
+            this.SetParamSimple(map, prefix + "EngineType", this.EngineType);
+            this.SetParamSimple(map, prefix + "TableType", this.TableType);
+            this.SetParamSimple(map, prefix + "ProjectId", this.ProjectId);
+            this.SetParamSimple(map, prefix + "ProjectName", this.ProjectName);
+            this.SetParamSimple(map, prefix + "ProjectDisplayName", this.ProjectDisplayName);
+            this.SetParamSimple(map, prefix + "TableOwnerId", this.TableOwnerId);
+            this.SetParamSimple(map, prefix + "TableOwnerName", this.TableOwnerName);
+            this.SetParamSimple(map, prefix + "StorageLocation", this.StorageLocation);
+            this.SetParamSimple(map, prefix + "Description", this.Description);
+            this.SetParamSimple(map, prefix + "IsPartitionTable", this.IsPartitionTable);
+            this.SetParamArraySimple(map, prefix + "PartitionColumns.", this.PartitionColumns);
+            this.SetParamSimple(map, prefix + "StorageFormat", this.StorageFormat);
+            this.SetParamSimple(map, prefix + "StorageSize", this.StorageSize);
+            this.SetParamSimple(map, prefix + "StorageSizeWithUnit", this.StorageSizeWithUnit);
+            this.SetParamSimple(map, prefix + "TotalSizeMb", this.TotalSizeMb);
+            this.SetParamSimple(map, prefix + "ReplicaCount", this.ReplicaCount);
+            this.SetParamSimple(map, prefix + "FileCount", this.FileCount);
+            this.SetParamSimple(map, prefix + "PartitionCount", this.PartitionCount);
+            this.SetParamSimple(map, prefix + "PartitionFieldCount", this.PartitionFieldCount);
+            this.SetParamSimple(map, prefix + "PartitionExpireDays", this.PartitionExpireDays);
+            this.SetParamSimple(map, prefix + "CreateTime", this.CreateTime);
+            this.SetParamSimple(map, prefix + "UpdateTime", this.UpdateTime);
+        }
+    }
+}
+